[(Power)TimerEntry] Refactor to use common base classes
Refactor TimerEntry.TimerEntry and PowerTimerEntry.TimerEntry to
put common code into TimerEntryBase.TimerEntry, and refactor
TimerEntry.TimerLog and PowerTimerEntry.TimerLog to put common code
into TimerEntryBase.TimerLog

In fact, TimerEntry.TimerLog and PowerTimerEntry.TimerLog contain
no code, but the class structure was retained for backward compatiblity
and to allow class overrides in the specific timer and power timer
cases if necessary.
