+++ lib/python/Screens/TimerEntry.py
@@ -1,22 +1,18 @@
 # -*- coding: utf-8 -*-
-from time import localtime, mktime, time, strftime
-from datetime import datetime
+from time import localtime, time, strftime
 
 from enigma import eEPGCache
 
 from Screens.Screen import Screen
 import ChannelSelection
 from ServiceReference import ServiceReference
-from Components.config import config, ConfigSelection, ConfigText, ConfigSubList, ConfigDateTime, ConfigClock, ConfigYesNo
+from Components.config import config, ConfigSelection, ConfigText, ConfigYesNo
 from Components.ActionMap import HelpableActionMap
-from Components.MenuList import MenuList
-from Components.Button import Button
 from Components.Label import Label
 from Components.Pixmap import Pixmap
 from Components.SystemInfo import SystemInfo
 from Components.UsageConfig import defaultMoviePath
-from Screens.Setup import Setup
-from Screens.HelpMenu import HelpableScreen
+from Screens.TimerEntryBase import TimerEntryBase, TimerLogBase
 from Screens.MovieSelection import getPreferredTagEditor
 from Screens.LocationBox import MovieLocationBox
 from Screens.ChoiceBox import ChoiceBox
@@ -25,32 +21,12 @@ from Screens.VirtualKeyBoard import VirtualKeyBoard
 from RecordTimer import AFTEREVENT
 
 
-class TimerEntry(Setup):
+class TimerEntry(TimerEntryBase):
 	def __init__(self, session, timer):
-		# Need to create some variables before Setup reads setup.xml
-		self.timer = timer
-		self.createConfig()
-
-		Setup.__init__(self, session=session, setup="timerentry")
-
-		# Are these actually skinned anywhere?
-		self["oktext"] = Label(_("Save"))
-		self["canceltext"] = Label(_("Cancel"))
-		self["ok"] = Pixmap()
-		self["cancel"] = Pixmap()
-
-		self["actions"] = HelpableActionMap(self, ["ConfigListActions", "GlobalActions", "PiPSetupActions"],
-		{
-			"save": (self.keySave, _("Save timer")),
-			"cancel": (self.keyCancel, _("Cancel timer creation / changes")),
-			"close": (self.keyCancel, _("Cancel timer creation / changes")),
-			"volumeUp": (self.incrementStart, _("Increment start time")),
-			"volumeDown": (self.decrementStart, _("Decrement start time")),
-			"size+": (self.incrementEnd, _("Increment end time")),
-			"size-": (self.decrementEnd, _("Decrement end time")),
-		}, prio=-1)
+		TimerEntryBase.__init__(self, session, timer, "timerentry")
 
 	def createConfig(self):
+		TimerEntryBase.createConfig(self)
 		justplay = self.timer.justplay
 		always_zap = self.timer.always_zap
 		pipzap = self.timer.pipzap
@@ -71,41 +47,6 @@ class TimerEntry(Setup):
 		elif self.timer.descramble:
 			recordingtype = "normal"
 
-		weekday_table = ("mon", "tue", "wed", "thu", "fri", "sat", "sun")
-
-		# calculate default values
-		day = []
-		weekday = 0
-		for x in (0, 1, 2, 3, 4, 5, 6):
-			day.append(0)
-		if self.timer.repeated: # repeated
-			type = "repeated"
-			if self.timer.repeated == 31: # Mon-Fri
-				repeated = "weekdays"
-			elif self.timer.repeated == 127: # daily
-				repeated = "daily"
-			else:
-				flags = self.timer.repeated
-				repeated = "user"
-				count = 0
-				for x in (0, 1, 2, 3, 4, 5, 6):
-					if flags == 1: # weekly
-# 						print "Set to weekday " + str(x)
-						weekday = x
-					if flags & 1 == 1: # set user defined flags
-						day[x] = 1
-						count += 1
-					else:
-						day[x] = 0
-					flags >>= 1
-				if count == 1:
-					repeated = "weekly"
-		else: # once
-			type = "once"
-			repeated = None
-			weekday = int(strftime("%u", localtime(self.timer.begin))) - 1
-			day[weekday] = 1
-
 		self.timerentry_justplay = ConfigSelection(choices = [
 			("zap", _("zap")), ("record", _("record")), ("zap+record", _("zap and record"))],
 			default = {0: "record", 1: "zap", 2: "zap+record"}[justplay + 2*always_zap])
@@ -115,7 +56,6 @@ class TimerEntry(Setup):
 			shutdownString = _("shut down")
 		self.timerentry_afterevent = ConfigSelection(choices = [("nothing", _("do nothing")), ("standby", _("go to standby")), ("deepstandby", shutdownString), ("auto", _("auto"))], default = afterevent)
 		self.timerentry_recordingtype = ConfigSelection(choices = [("normal", _("normal")), ("descrambled+ecm", _("descramble and record ecm")), ("scrambled+ecm", _("don't descramble, record ecm"))], default = recordingtype)
-		self.timerentry_type = ConfigSelection(choices = [("once",_("once")), ("repeated", _("repeated"))], default = type)
 		self.timerentry_name = ConfigText(default = self.timer.name.replace('\xc2\x86', '').replace('\xc2\x87', '').encode("utf-8"), visible_width = 50, fixed_size = False)
 		self.timerentry_description = ConfigText(default = self.timer.description, visible_width = 50, fixed_size = False)
 		self.timerentry_tags = self.timer.tags[:]
@@ -128,15 +68,11 @@ class TimerEntry(Setup):
 
 		self.timerentry_tagsset = ConfigSelection(choices = [not self.timerentry_tags and "None" or " ".join(self.timerentry_tags)])
 
-		self.timerentry_repeated = ConfigSelection(default = repeated, choices = [("weekly", _("weekly")), ("daily", _("daily")), ("weekdays", _("Mon-Fri")), ("user", _("user defined"))])
 		self.timerentry_renamerepeat = ConfigYesNo(default = rename_repeat)
 
 		self.timerentry_pipzap = ConfigYesNo(default = pipzap)
 		self.timerentry_conflictdetection = ConfigYesNo(default = conflict_detection)
 
-		self.timerentry_date = ConfigDateTime(default = self.timer.begin, formatstring = config.usage.date.full.value, increment = 86400)
-		self.timerentry_starttime = ConfigClock(default = self.timer.begin)
-		self.timerentry_endtime = ConfigClock(default = self.timer.end)
 		self.timerentry_showendtime = ConfigSelection(default = False, choices = [(True, _("yes")), (False, _("no"))])
 
 		default = self.timer.dirname or defaultMoviePath()
@@ -144,14 +80,6 @@ class TimerEntry(Setup):
 		if default not in tmp:
 			tmp.append(default)
 		self.timerentry_dirname = ConfigSelection(default = default, choices = tmp)
-
-		self.timerentry_repeatedbegindate = ConfigDateTime(default = self.timer.repeatedbegindate, formatstring = config.usage.date.full.value, increment = 86400)
-
-		self.timerentry_weekday = ConfigSelection(default = weekday_table[weekday], choices = [("mon",_("Monday")), ("tue", _("Tuesday")), ("wed",_("Wednesday")), ("thu", _("Thursday")), ("fri", _("Friday")), ("sat", _("Saturday")), ("sun", _("Sunday"))])
-
-		self.timerentry_day = ConfigSubList()
-		for x in (0, 1, 2, 3, 4, 5, 6):
-			self.timerentry_day.append(ConfigYesNo(default = day[x]))
 
 		# FIXME some service-chooser needed here
 		servicename = "N/A"
@@ -232,22 +160,8 @@ class TimerEntry(Setup):
 			self.timerentry_service.setCurrentText(self.timerentry_service_ref.getServiceName())
 			self.invalidateConfigEntry(self.timerentry_service)
 
-	def getTimestamp(self, date, mytime):
-		d = localtime(date)
-		dt = datetime(d.tm_year, d.tm_mon, d.tm_mday, mytime[0], mytime[1])
-		return int(mktime(dt.timetuple()))
-
 	def getBeginEnd(self):
-		date = self.timerentry_date.value
-		endtime = self.timerentry_endtime.value
-		starttime = self.timerentry_starttime.value
-
-		begin = self.getTimestamp(date, starttime)
-		end = self.getTimestamp(date, endtime)
-
-		# if the endtime is less than the starttime, add 1 day.
-		if end < begin:
-			end += 86400
+		begin, end = TimerEntryBase.getBeginEnd(self)
 
 		# if the timer type is a Zap and no end is set, set duration to 1 second so time is shown in EPG's.
 		if self.timerentry_justplay.value == "zap":
@@ -268,10 +182,6 @@ class TimerEntry(Setup):
 			self.finishedChannelSelection(*args)
 			self.keySave()
 
-	def keyGo(self, result = None):
-		print "[TimerEntry] keyGo() is deprecated, call keySave() instead"
-		self.keySave(result)
-
 	def keySave(self, result = None):
 		if not self.timerentry_service_ref.isRecordable():
 			self.session.openWithCallback(self.selectChannelSelector, MessageBox, _("You didn't select a channel to record from."), MessageBox.TYPE_ERROR)
@@ -286,7 +196,6 @@ class TimerEntry(Setup):
 		if self.timerentry_justplay.value == "zap":
 			if not self.timerentry_showendtime.value:
 				self.timerentry_endtime.value = self.timerentry_starttime.value
-		self.timer.resetRepeated()
 		self.timer.afterEvent = {
 			"nothing": AFTEREVENT.NONE,
 			"deepstandby": AFTEREVENT.DEEPSTANDBY,
@@ -320,34 +229,8 @@ class TimerEntry(Setup):
 
 		if self.timerentry_type.value == "once":
 			self.timer.begin, self.timer.end = self.getBeginEnd()
-		if self.timerentry_type.value == "repeated":
-			if self.timerentry_repeated.value == "daily":
-				for x in (0, 1, 2, 3, 4, 5, 6):
-					self.timer.setRepeated(x)
 
-			if self.timerentry_repeated.value == "weekly":
-				self.timer.setRepeated(self.timerentry_weekday.index)
-
-			if self.timerentry_repeated.value == "weekdays":
-				for x in (0, 1, 2, 3, 4):
-					self.timer.setRepeated(x)
-
-			if self.timerentry_repeated.value == "user":
-				for x in (0, 1, 2, 3, 4, 5, 6):
-					if self.timerentry_day[x].value:
-						self.timer.setRepeated(x)
-
-			self.timer.repeatedbegindate = self.getTimestamp(self.timerentry_repeatedbegindate.value, self.timerentry_starttime.value)
-			if self.timer.repeated:
-				self.timer.begin = self.getTimestamp(self.timerentry_repeatedbegindate.value, self.timerentry_starttime.value)
-				self.timer.end = self.getTimestamp(self.timerentry_repeatedbegindate.value, self.timerentry_endtime.value)
-			else:
-				self.timer.begin = self.getTimestamp(time(), self.timerentry_starttime.value)
-				self.timer.end = self.getTimestamp(time(), self.timerentry_endtime.value)
-
-			# when a timer end is set before the start, add 1 day
-			if self.timer.end < self.timer.begin:
-				self.timer.end += 86400
+		TimerEntryBase.keySave(self)
 
 		if self.timer.eit is not None:
 			event = eEPGCache.getInstance().lookupEventId(self.timer.service_ref.ref, self.timer.eit)
@@ -368,6 +251,7 @@ class TimerEntry(Setup):
 				elif n > 0:
 					parent = self.timer.service_ref.ref
 					self.timer.service_ref = ServiceReference(event.getLinkageService(parent, 0))
+
 		self.saveTimer()
 		self.close((True, self.timer))
 
@@ -375,29 +259,13 @@ class TimerEntry(Setup):
 		self.timerentry_justplay.selectNext()
 		self.invalidateConfigEntry(self.timerentry_justplay)
 
-	def incrementStart(self):
-		self.timerentry_starttime.increment()
-		self.invalidateConfigEntry(self.timerentry_starttime)
-		if self.timerentry_type.value == "once" and self.timerentry_starttime.value == [0, 0]:
-			self.timerentry_date.value += 86400
-			self.invalidateConfigEntry(self.timerentry_date)
-
-	def decrementStart(self):
-		self.timerentry_starttime.decrement()
-		self.invalidateConfigEntry(self.timerentry_starttime)
-		if self.timerentry_type.value == "once" and self.timerentry_starttime.value == [23, 59]:
-			self.timerentry_date.value -= 86400
-			self.invalidateConfigEntry(self.timerentry_date)
-
 	def incrementEnd(self):
-		if self.timerentry_justplay.value != "zap" or self.timerentry_showendtime.value:
-			self.timerentry_endtime.increment()
-			self.invalidateConfigEntry(self.timerentry_endtime)
+		if self.timerentry_showendtime.value or self.timerentry_justplay.value != "zap":
+			TimerEntryBase.incrementEnd(self)
 
 	def decrementEnd(self):
-		if self.timerentry_justplay.value != "zap" or self.timerentry_showendtime.value:
-			self.timerentry_endtime.decrement()
-			self.invalidateConfigEntry(self.timerentry_endtime)
+		if self.timerentry_showendtime.value or self.timerentry_justplay.value != "zap":
+			TimerEntryBase.decrementEnd(self)
 
 	def subserviceSelected(self, service):
 		if not service is None:
@@ -407,9 +275,6 @@ class TimerEntry(Setup):
 
 	def saveTimer(self):
 		self.session.nav.RecordTimer.saveTimer()
-
-	def keyCancel(self):
-		self.closeConfigList(((False,),))
 
 	def pathSelected(self, res):
 		if res is not None:
@@ -423,100 +288,8 @@ class TimerEntry(Setup):
 			self.timerentry_tagsset.setChoices([not ret and "None" or " ".join(ret)])
 			self.invalidateConfigEntry(self.timerentry_tagsset)
 
-	def invalidateConfigEntry(self, conf):
-		for ent in self.list:
-			if ent[1] is conf:
-				self["config"].invalidate(ent)
-
-class TimerLog(Screen, HelpableScreen):
-	def __init__(self, session, timer):
-		Screen.__init__(self, session)
-		HelpableScreen.__init__(self)
-		self.setTitle(_("Log"))
-
-		self.timer = timer
-		self.log_entries = self.timer.log_entries[:]
-
-		self.fillLogList()
-
-		self["loglist"] = MenuList(self.list)
-		self["logentry"] = Label()
-
-		self["key_red"] = Button(_("Delete entry"))
-		self["key_blue"] = Button(_("Clear log"))
-
-		self["loglist"].onSelectionChanged.append(self.updateText)
-
-		self["actions"] = HelpableActionMap(self, ["OkCancelActions", "DirectionActions", "ColorActions"],
-		{
-			"ok": (self.keyClose, _("Close screen")),
-			"cancel": (self.keyClose, _("Close screen")),
-			"up": (self.moveUp, _("Move up a line")),
-			"down": (self.moveDown, _("Move down a line")),
-			"left": (self.pageUp, _("Move up a screen")),
-			"right": (self.pageDown, _("Move down a screen")),
-			"red": (self.deleteEntry, _("Delete log entry")),
-			"blue": (self.clearLog, _("Delete all log entries")),
-		})
-
-	def deleteEntry(self):
-		cur = self["loglist"].getCurrent()
-		if cur is None:
-			return
-		self.log_entries.remove(cur[1])
-		self.fillLogList()
-		self["loglist"].l.setList(self.list)
-		self.updateText()
-
-	def fillLogList(self):
-		self.list = [(str(strftime(config.usage.date.daylong.value + " " + config.usage.time.short.value, localtime(x[0])) + " - " + x[2]), x) for x in self.log_entries]
-
-	def clearLog(self):
-		self.log_entries = []
-		self.fillLogList()
-		self["loglist"].l.setList(self.list)
-		self.updateText()
-
-	def keyClose(self):
-		if self.timer.log_entries != self.log_entries:
-			self.timer.log_entries = self.log_entries
-			self.close((True, self.timer))
-		else:
-			self.close((False,))
-
-	def up(self):
-		print "[TimerLog] up() is deprecated, call moveUp() instead"
-		self.moveUp()
-
-	def moveUp(self):
-		self["loglist"].moveUp()
-
-	def down(self):
-		print "[TimerLog] down() is deprecated, call moveDown() instead"
-		self.moveDown()
-
-	def moveDown(self):
-		self["loglist"].moveDown()
-
-	def left(self):
-		print "[TimerLog] left() is deprecated, call pageUp() instead"
-		self.pageUp()
-
-	def pageUp(self):
-		self["loglist"].pageUp()
-
-	def right(self):
-		print "[TimerLog] right() is deprecated, call pageDown() instead"
-		self.pageDown()
-
-	def pageDown(self):
-		self["loglist"].pageDown()
-
-	def updateText(self):
-		if self.list:
-			self["logentry"].text = str(self["loglist"].getCurrent()[1][2])
-		else:
-			self["logentry"].text = ""
+class TimerLog(TimerLogBase):
+	pass
 
 class InstantRecordTimerEntry(TimerEntry):
 	def __init__(self, session, timer, zap):
